fix: Linking out from Dianomi ads within a WebView
